- Update Xen patches to 3.16.6 and c/s 1266.
suse-commit: cfbe55c1df9a7c4ce5efee06f463a83f8cc9f892

+++ include/xen/evtchn.h
@@ -55,7 +55,7 @@ struct irq_cfg {
 	union {
 		int bindcount; /* for dynamic IRQs */
 #ifdef CONFIG_X86_IO_APIC
-		u8 vector; /* for physical IRQs */
+		unsigned int vector; /* for physical IRQs */
 #endif
 	};
 };
